+++ main/src/ca/uwaterloo/flix/util/Options.scala
@@ -16,6 +16,8 @@
 
 package ca.uwaterloo.flix.util
 
+import scala.concurrent.duration.Duration
+
 object Options {
   /**
     * Default options.
@@ -25,6 +27,7 @@ object Options {
     evaluation = Evaluation.Compiled,
     monitor = false,
     optimize = false,
+    timeout = Duration.Inf,
     threads = Runtime.getRuntime.availableProcessors(),
     verbosity = Verbosity.Normal,
     verifier = false
@@ -41,13 +44,14 @@ object Options {
   *
   * @param debug      enables the emission of debugging information.
   * @param evaluation selects the evaluation strategy,
-  * @param monitor    enables the debugger and profiler.
   * @param optimize   enables compiler optimizations.
+  * @param monitor    enables the debugger and profiler.
+  * @param timeout    selects the solver timeout.
   * @param threads    selects the number of threads to use.
   * @param verbosity  selects the level of verbosity.
   * @param verifier   enables the verifier.
   */
-case class Options(debug: Boolean, evaluation: Evaluation, monitor: Boolean, optimize: Boolean, threads: Int, verbosity: Verbosity, verifier: Boolean)
+case class Options(debug: Boolean, evaluation: Evaluation, optimize: Boolean, monitor: Boolean, timeout: Duration, threads: Int, verbosity: Verbosity, verifier: Boolean)
 
 /**
   * An option to control the level of verbosity.