+++ main/test/ca/uwaterloo/flix/language/phase/TestParser.scala
@@ -2089,6 +2089,34 @@ class TestParser extends FunSuite {
     new Flix().addStr(input).compile().get
   }
 
+  test("Declaration.Fact.Head.True") {
+    val input = "true."
+    new Flix().addStr(input).compile().get
+  }
+
+  test("Declaration.Fact.Head.False") {
+    val input = "false."
+    new Flix().addStr(input).compile().get
+  }
+
+  test("Declaration.Rule.Head.True") {
+    val input =
+      """rel R(a: Int, b: Int)
+        |
+        |true :- R(x, y).
+      """.stripMargin
+    new Flix().addStr(input).compile().get
+  }
+
+  test("Declaration.Rule.Head.False") {
+    val input =
+      """rel R(a: Int, b: Int)
+        |
+        |false :- R(x, y).
+      """.stripMargin
+    new Flix().addStr(input).compile().get
+  }
+
   test("Declaration.Rule.04") {
     val input =
       """def f: Int = 42