+++ CHANGELOG.md
@@ -1,6 +1,10 @@
+# Changelog
+
 HEAD:
+    - Added delta debugging (see --delta and --tutorial delta-debugging).
     - Added parallel rule evaluation. (Use --threads to control the number of concurrent threads).
-    - Added tutorial. (Run --tutorial <file> to output the tutorial to a file).
+    - Added tutorial. (see --tutorial introduction).
+    - Added --timeout parameter.
     - Improved command line parsing. Added a --help flag to display available options.
 
 2016-06-08: